create common and add styles for commons
