+++ src/common/styles.js
@@ -1,0 +1,11 @@
+import { Link } from 'react-router-dom'
+import styled from 'styled-components'
+
+export const LinkStyled = styled(Link)`
+ font-size: 1rem;
+ padding-left: 2rem ;
+ svg {
+   color: var(--blue-02);
+   margin-right: .3rem ;
+ }
+`
