+++ src/pages/Activities/Activity/styles.js
@@ -1,16 +1,6 @@
-import { Link } from 'react-router-dom'
 import styled from 'styled-components'
 
 export const Container = styled.div`
   padding-top: 1rem;
   text-align: left;  
-`
-
-export const LinkStyled = styled(Link)`
- font-size: 1rem;
- padding-left: 2rem ;
- svg {
-   color: var(--blue-02);
-   margin-right: .3rem ;
- }
 `
